+++ ProjectModernpp/Server/Server/Point.cpp
@@ -2,12 +2,12 @@ import Point;
 
 #include <utility> 
 
-Point::Point(int x, int y)
+Point::Point(int8_t x, int8_t y)
 	: m_x{ x }, m_y{ y }, m_data{ nullptr } {}
 
 
 Point::Point(const Point& other)
-	: m_x(other.m_x), m_y(other.m_y), m_data(new int[2] {other.m_x, other.m_y}) {}
+	: m_x(other.m_x), m_y(other.m_y), m_data(new int8_t[2]{ other.m_x, other.m_y }) {}
 
 Point::Point(Point&& other) noexcept
 	: m_x(other.m_x), m_y(other.m_y), m_data(other.m_data) {
@@ -42,7 +42,7 @@ Point& Point::operator=(const Point& position)
 	m_y = position.m_y;
 
 	if (position.m_data) {
-		m_data = new int(*position.m_data);
+		m_data = new int8_t(*position.m_data);
 	}
 	else {
 		m_data = nullptr;
@@ -76,12 +76,12 @@ bool Point::operator!=(const Point& other) const
 	return !(*this == other);
 }
 
-size_t Point::GetX() const
+int8_t Point::GetX() const
 {
 	return m_x;
 }
 
-size_t Point::GetY() const
+int8_t Point::GetY() const
 {
 	return m_y;
 }
